Move ConstEvaluator into simplity_expressions.rs
